moved config to akiradocs.config.json
