+++ packages/akiradocs/akiradocs.config.json
@@ -10,7 +10,7 @@
             "show": true
         },
         "title": {
-            "text": "",
+            "text": "Akira Docs",
             "show": false
         },
         "description": {
@@ -25,7 +25,7 @@
             "height": 30
         },
         "title": {
-            "text": "",
+            "text": "Akira Docs",
             "show": false
         },
         "searchPlaceholder": "Search Akira...",
@@ -41,10 +41,6 @@
             { "name": "GitHub", "url": "https://github.com/Cloud-Code-AI/akiradocs-template", "icon": "/github.svg" },
             { "name": "Twitter", "url": "https://x.com/CloudCodeAI", "icon": "/twitter.svg" },
             { "name": "LinkedIn", "url": "https://www.linkedin.com/company/cloudcode-ai", "icon": "/linkedin.svg" }
-        ],
-        "madeWithLove": {
-            "show": true,
-            "team": "Akira Docs"
-        }
+        ]
     }
 }
