+++ packages/akiradocs/src/app/[type]/[...slug]/page.tsx
@@ -12,7 +12,7 @@ import { getFooterConfig } from '@/lib/footerConfig'
 import { Button } from 'akiradocs-ui'
 import { Edit2 } from 'lucide-react'
 import { PageBreadcrumb } from 'akiradocs-ui'
-import { getNextPrevPages } from '@/utils/navigationUtils'
+import { getNextPrevPages } from '@/lib/navigationUtils'
 import { PageNavigation } from 'akiradocs-ui'
 import { MainTitle, SubTitle } from 'akiradocs-ui'
 import { SEO } from 'akiradocs-ui'
