+++ packages/akiradocs/src/app/aiSearch/page.tsx
@@ -11,14 +11,14 @@ import { getRecommendedArticles } from '@/lib/recommendedArticles'
 import { getSearchConfig } from '@/lib/searchConfig'
 import { Button } from 'akiradocs-ui'
 import Link from 'next/link'
-
+import { getAkiradocsConfig } from '@/lib/getAkiradocsConfig'
 
 export default function Home() {
   const [query, setQuery] = useState('')
   const [aiResponse, setAiResponse] = useState('')
   const recommendedArticles = getRecommendedArticles()
   const searchConfig = getSearchConfig()
-  const config = require('../../../_contents/_config.json')
+  const config = getAkiradocsConfig()
 
   // If AI Search is disabled, show the disabled message
   if (!config.aiSearch) {
