+++ packages/akiradocs/src/components/content/layout/Footer.tsx
@@ -18,10 +18,6 @@ interface SocialLink {
 interface FooterProps {
   companyName?: string;
   socialLinks?: SocialLink[];
-  madeWithLove?: {
-    show?: boolean;
-    team?: string;
-  };
 }
 
 const defaultProps = {
@@ -32,17 +28,12 @@ const defaultProps = {
       url: "https://github.com",
       icon: "/github.svg"
     }
-  ],
-  madeWithLove: {
-    show: true,
-    team: "Development"
-  }
+  ]
 } as const;
 
 export const Footer: React.FC<FooterProps> = ({ 
   companyName = defaultProps.companyName,
   socialLinks = defaultProps.socialLinks,
-  madeWithLove = defaultProps.madeWithLove
 }) => {
   const currentYear = new Date().getFullYear()
 
@@ -58,12 +49,7 @@ export const Footer: React.FC<FooterProps> = ({
           <div className="flex flex-col md:flex-row justify-between items-center space-y-4 md:space-y-0">
             <div className="flex items-center space-x-2 text-sm text-muted-foreground">
               <span>© {currentYear} {companyName}. All rights reserved.</span>
-              {madeWithLove.show && (
-                <>
-                  <span>|</span>
-                  <span>Made with <Heart className="inline-block w-4 h-4 text-red-500" /> by the {madeWithLove.team} team</span>
-                </>
-              )}
+              <span>Made with <Heart className="inline-block w-4 h-4 text-red-500" /> by the Akiradocs team</span>
             </div>
             <div className="flex space-x-4">
               {socialLinks.map((link, index) => (
