+++ packages/akiradocs/src/components/content/renderers/BlockRenderer.tsx
@@ -5,7 +5,7 @@ import { HeadingTitle } from 'akiradocs-ui'
 import { List } from 'akiradocs-ui'
 import { Blockquote } from 'akiradocs-ui'
 import { Divider } from 'akiradocs-ui'
-import { CodeBlock } from '@/components/content/blocks/CodeBlock'
+import { CodeBlock } from 'akiradocs-ui'
 // import { Image } from '../blocks/Image'
 // import { Table } from '../blocks/Table'
 // import { ToggleList } from '../blocks/ToggleList'
