+++ packages/akiradocs/src/lib/footerConfig.ts
@@ -1,3 +1,4 @@
+  import { getAkiradocsConfig } from "./getAkiradocsConfig";
   interface SocialLink {
     name: string;
     url: string;
@@ -7,10 +8,6 @@
   type FooterConfig = {
     companyName: string;
     socialLinks: SocialLink[];
-    madeWithLove: {
-      show: boolean;
-      team: string;
-    };
   };
   declare var require: {
     context(
@@ -21,15 +18,12 @@
   };
   export function getFooterConfig(): FooterConfig {
     // Using require.context to get the config file
-    const context = require.context('../../_contents', false, /_config\.json$/);
-    const configPath = context.keys()[0];
-    const config = context(configPath);
+    const config = getAkiradocsConfig();
   
     // Return only the header-related config
     return {
         companyName: config.footer?.companyName,
         socialLinks: config.footer?.socialLinks,
-        madeWithLove: config.footer?.madeWithLove,
     };
   }
