+++ packages/akiradocs/src/lib/getAkiradocsConfig.ts
@@ -1,0 +1,5 @@
+import config from '../../akiradocs.config.json';
+
+export function getAkiradocsConfig(): any {
+    return config;
+}
