--- packages/akiradocs/src/lib/getNavigation.ts
@@ -1,1 +1,0 @@
-"use client"
