+++ packages/akiradocs/src/lib/searchConfig.ts
@@ -1,3 +1,5 @@
+import { getAkiradocsConfig } from "./getAkiradocsConfig";
+
 type SearchConfig = {
     logo?: {
         path: string;
@@ -15,19 +17,9 @@ type SearchConfig = {
     };
 };
 
-declare var require: {
-  context(
-    directory: string,
-    useSubdirectories: boolean,
-    regExp: RegExp
-  ): any;
-};
 
 export function getSearchConfig(): SearchConfig {
-    const context = require.context('../../_contents', false, /_config\.json$/);
-    const configPath = context.keys()[0];
-    const config = context(configPath);
-
+    const config = getAkiradocsConfig();
     return {
         logo: config.search?.logo,
         title: config.search?.title,
