+++ packages/akiradocs/tsconfig.json
@@ -25,6 +25,12 @@
     "paths": {
       "@/*": [
         "./src/*"
+      ],
+      "@/config": [
+        "../akiradocs.config.json"
+      ],
+      "@/compiled": [
+        "compiled/**/*.json"
       ]
     },
     "target": "ES2017"
@@ -34,7 +40,10 @@
     "**/*.ts",
     "**/*.tsx",
     ".next/types/**/*.ts",
-    "content/**/*.json"
+    "content/**/*.json",
+    "../akiradocs.config.json",
+    "compiled/**/*.json"
+
   ],
   "exclude": [
     "node_modules"
